+++ osu.Game/GameModes/Backgrounds/BackgroundModeDefault.cs
@@ -1,0 +1,17 @@
+﻿//Copyright (c) 2007-2016 ppy Pty Ltd <contact@ppy.sh>.
+//Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Game.Graphics.Background;
+
+namespace osu.Game.GameModes.Backgrounds
+{
+    public class BackgroundModeDefault : BackgroundMode
+    {
+        public override void Load()
+        {
+            base.Load();
+
+            Add(new Background());
+        }
+    }
+}
