+++ osu.Game/GameModes/Backgrounds/BackgroundModeEmpty.cs
@@ -1,0 +1,10 @@
+﻿//Copyright (c) 2007-2016 ppy Pty Ltd <contact@ppy.sh>.
+//Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+namespace osu.Game.GameModes.Backgrounds
+{
+    public class BackgroundModeEmpty : BackgroundMode
+    {
+
+    }
+}
