+++ osu.Game/GameModes/Multiplayer/Match.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using osu.Framework.GameModes;
+using osu.Game.GameModes.Backgrounds;
 using osu.Game.GameModes.Play;
 using OpenTK.Graphics;
