+++ osu.Game/GameModes/Play/Results.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using osu.Framework.GameModes;
+using osu.Game.GameModes.Backgrounds;
 using OpenTK.Graphics;
 
 namespace osu.Game.GameModes.Play
